ANDROID: Add untag hacks to inet_release function
To prevent protential risk of memory leak caused by closing socket with
out untag it from qtaguid module, the qtaguid module now do not hold any
socket file reference count. Instead, it will increase the sk_refcnt of
the sk struct to prevent a reuse of the socket pointer.  And when a socket
is released. It will delete the tag if the socket is previously tagged so
no more resources is held by xt_qtaguid moudle.

Bug: 36374484
Test: compile and run test under system/extra/test/iptables

Change-Id: I40223b9970c89274c3b5bd339d27e1d013492421

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>